+++ NEWS.md
@@ -1,4 +1,8 @@
-# clootl v1.0.0
+# clootl v0.1.2
+- Updates data to version 1.5, containing 2024 taxonomy and tree version 1.5
+
+# clootl v0.1.1
 - Initial CRAN submission.
-- Tree and Taxonomy data is available as an R data object named ...
-- updated Tree and taxonomy data is available via download with function...
+- Default Tree and Taxonomy data is available as an R data object named `clootl_data`
+- Updated Tree and taxonomy data is available via download with function `get_avesdata_repo()`
+- Updated tree in clootl_data.rda v1.4